tokenize chats
