+++ rigging/__init__.py
@@ -47,6 +47,16 @@ from rigging.message import (
 )
 from rigging.model import Model, attr, element, wrapped
 from rigging.prompt import Ctx, Prompt, prompt
+from rigging.tokenize import (
+    ChatFormatter,
+    Decoder,
+    Encoder,
+    TokenizedChat,
+    Tokenizer,
+    TokenSlice,
+    find_in_tokens,
+    get_tokenizer,
+)
 from rigging.tools import Tool, mcp, robopages, tool, tool_method
 from rigging.util import await_
 from rigging.version import VERSION
@@ -55,6 +65,7 @@ __version__ = VERSION
 
 __all__ = [
     "Chat",
+    "ChatFormatter",
     "ChatPipeline",
     "Completion",
     "CompletionPipeline",
@@ -62,6 +73,8 @@ __all__ = [
     "ContentImageUrl",
     "ContentText",
     "Ctx",
+    "Decoder",
+    "Encoder",
     "GenerateParams",
     "GeneratedMessage",
     "GeneratedText",
@@ -79,6 +92,9 @@ __all__ = [
     "Stop",
     "ThenChatCallback",
     "ThenCompletionCallback",
+    "TokenSlice",
+    "TokenizedChat",
+    "Tokenizer",
     "Tool",
     "attr",
     "await_",
@@ -87,8 +103,10 @@ __all__ = [
     "data",
     "element",
     "error",
+    "find_in_tokens",
     "generator",
     "get_generator",
+    "get_tokenizer",
     "interact",
     "logging",
     "mcp",
